Recovered C9~C11; in the process of recreatingvias
This is an re-edit on the March 22nd version. Have finished recreating the 3 sets of vias (6 vias) on the right side SMA connectors.
We use: 16.4 mils for top layer wires and 17mils for bottom layer wires.
C8-C11 are expected to be connected identically as are C4-C7 to generate no errors in connectivity check.